requirements.txt
google-generativeai==0.5.2
pillow==10.3.0
streamlit==1.33.0
streamlit-option-menu==0.3.12
